Delete scripts for old API
