+++ tests/unit/test_cqc_api.py
@@ -2,7 +2,7 @@ import unittest
 from unittest.mock import Mock, patch, call
 from typing import Generator
 
-from utils import cqc_api_new as cqc
+from utils import cqc_api as cqc
 
 
 class CqcApiTests(unittest.TestCase):
