+++ utils/cqc_api.py
@@ -1,45 +1,57 @@
 from ratelimit import limits, sleep_and_retry
-from time import sleep
+import time
 
 import requests
 
 CQC_API_VERSION = "v1"
-RATE_LIMIT = (
-    400  # Max ratelimit = 600 per minute without partnercode, 2000 with a partnercode
-)
+RATE_LIMIT = 400
 ONE_MINUTE = 60
 DEFAULT_PAGE_SIZE = 500
+CQC_API_BASE_URL = "https://api.service.cqc.org.uk"
+USER_AGENT = "SkillsForCare"
 
 
 @sleep_and_retry
 @limits(calls=RATE_LIMIT, period=ONE_MINUTE)
-def call_api(url, query_params=None):
-    response = requests.get(url, query_params)
+def call_api(url, query_params=None, headers_dict=None):
+    response = requests.get(url, query_params, headers=headers_dict)
 
     while response.status_code == 429:
         print("Sleeping for ten seconds due to rate limiting")
-        sleep(10)
-        response = requests.get(url, query_params)
+        time.sleep(10)
+        response = requests.get(url, query_params, headers=headers_dict)
 
-    if response.status_code != 200:
+    if (response.status_code == 403) & (headers_dict is None):
+        raise Exception(
+            "API response: {}, ensure you have set a User-Agent header".format(
+                response.status_code
+            )
+        )
+
+    if response.status_code is not 200:
         raise Exception("API response: {}".format(response.status_code))
 
     return response.json()
 
 
 def get_all_objects(
-    stream,
-    object_type,
-    object_identifier,
-    partner_code,
+    object_type: str,
+    object_identifier: str,
+    cqc_api_primary_key: str,
     per_page=DEFAULT_PAGE_SIZE,
 ):
-    url = f"https://api.cqc.org.uk/public/{CQC_API_VERSION}/{object_type}"
+    url = f"{CQC_API_BASE_URL}/public/{CQC_API_VERSION}/{object_type}"
 
-    total_pages = call_api(url, {"perPage": per_page, "partnerCode": partner_code})[
-        "totalPages"
-    ]
-    all_objects = []
+    total_pages = call_api(
+        url,
+        {
+            "perPage": per_page,
+        },
+        headers_dict={
+            "User-Agent": USER_AGENT,
+            "Ocp-Apim-Subscription-Key": cqc_api_primary_key,
+        },
+    )["totalPages"]
 
     print(f"Total pages: {total_pages}")
     print(f"Beginning CQC bulk download of {object_type}...")
@@ -47,16 +59,10 @@ def get_all_objects(
     for page_number in range(1, total_pages + 1):
         print(f"Collecting {object_type} from API page {page_number}/{total_pages}")
         page_locations = get_page_objects(
-            url, page_number, object_type, object_identifier, partner_code
+            url, page_number, object_type, object_identifier, cqc_api_primary_key
         )
 
-        if stream:
-            yield page_locations
-        else:
-            all_objects.append(page_locations)
-
-    if not stream:
-        return all_objects
+        yield page_locations
 
 
 def get_page_objects(
@@ -64,25 +70,38 @@ def get_page_objects(
     page_number,
     object_type,
     object_identifier,
-    partner_code,
+    cqc_api_primary_key,
     per_page=DEFAULT_PAGE_SIZE,
 ):
     page_objects = []
     response_body = call_api(
-        url, {"page": page_number, "perPage": per_page, "partnerCode": partner_code}
+        url,
+        {"page": page_number, "perPage": per_page},
+        headers_dict={
+            "User-Agent": USER_AGENT,
+            "Ocp-Apim-Subscription-Key": cqc_api_primary_key,
+        },
     )
 
     for resource in response_body[object_type]:
         returned_object = get_object(
-            resource[object_identifier], object_type, partner_code
+            resource[object_identifier], object_type, cqc_api_primary_key
         )
         page_objects.append(returned_object)
 
     return page_objects
 
 
-def get_object(cqc_location_id, object_type, partner_code):
-    url = f"https://api.cqc.org.uk/public/{CQC_API_VERSION}/{object_type}/"
+def get_object(cqc_location_id, object_type, cqc_api_primary_key):
+    url = f"{CQC_API_BASE_URL}/public/{CQC_API_VERSION}/{object_type}/"
 
-    location_body = call_api(url + cqc_location_id, {"partnerCode": partner_code})
-    return location_body
+    response = call_api(
+        url + cqc_location_id,
+        query_params=None,  # This endpoint no longer supports a Partner Code
+        headers_dict={
+            "User-Agent": USER_AGENT,
+            "Ocp-Apim-Subscription-Key": cqc_api_primary_key,
+        },
+    )
+
+    return response
